requirements.txt
pyautogui
pillow
numpy
tensorflow
matplotlib
